Conflicts:
	Projects/Gameplay/Src/ServerUpdateSystem.cpp

+++ Projects/Gameplay/Src/GameStatsSystem.cpp
@@ -64,7 +64,7 @@ void GameStatsSystem::updateStats( const UpdateClientStatsPacket* p_packet )
 		PlayerStats stats;
 		stats.name	= toString(p_packet->playerIdentities[i]);
 		stats.score = p_packet->scores[i];
-		stats.ping	= p_packet->ping[i];
+		stats.ping	= static_cast<int>(p_packet->ping[i]);
 
 		m_infoPanel->updateRow(i, stats);
 	}
